add recursive control

+++ src/find_it.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int find_it(int *array, int size) {
+int find_it(int *array, unsigned long long size) {
     if (array == NULL || size <= 0) {
         fprintf(stderr, "error: wrong data.\n");
         exit(EXIT_FAILURE);
